Finished RocketInterface.
Added support to index buffers and draw element arrays.

VertexBuffer is now a templatized HardwareBuffer and we have to "specializations" VertexBuffer and ElementBuffer that are the same thing, just for abstraction.

The libRocket context management and input handling is application responsability, we only provide a libRocket interface to Seed system, no more integration will be done (as with physics engines like Box2D and Bullet).